requirements.txt
setuptools>=41.0.0
Flask==1.1.2
flask_wtf
numpy==1.18.2
pandas==1.0.3
nltk==3.4.5
Keras==2.3.1
requests==2.23.0
pygame==1.9.6
COVID19Py==0.3.0
gunicorn==19.9.0
billboard.py
pytz==2019.3
tensorflow==2.2.0

